task_2/requirements.txt
allure-pytest==2.13.2
allure-python-commons==2.13.2
page-objects==1.1.0
pytest==7.4.2
requests==2.31.0
selenium==4.14.0